chore: move types into fred-types
