+++ fred-types/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "fred-types"
+version = "2.1.3"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+redis-protocol = "0.1.2"
